+++ packages/client/src/constants/toast.ts
@@ -1,5 +1,5 @@
-export const VOTE_START = '📢 투표가 시작됩니다!';
-export const VOTE_ALARM = '📢 투표가 10초 남았습니다!';
+export const VOTE_START = '📢 투표가 시작됩니다.';
+export const VOTE_ALARM = '📢 투표가 10초 남았습니다.';
 export const VOTE_END = '📢 투표가 종료되었습니다.';
-export const NIGHT = '🌒 밤이 되었습니다... 개인 능력을 사용해주세요.';
-export const DAY = '☀️ 낮이 되었습니다... 투표로 희생될 사람을 결정해주세요.';
+export const NIGHT = '🌒 밤이 되었습니다.';
+export const DAY = '☀️ 낮이 되었습니다.';
